chore: move helper functions to methods on relative structs, other cleanup
