+++ internal/config/config.go
@@ -69,8 +69,10 @@ type Query struct {
 	// Names is a list of TypeName entries that will be found at the above Path.
 	Endpoints []EndpointConfig `yaml:"endpoints,omitempty"`
 }
+
 type Command struct {
 	Name              string        `yaml:"name,omitempty"`
+	FileName          string        `yaml:"fileName,omitempty"`
 	ShortDescription  string        `yaml:"shortDescription,omitempty"`
 	LongDescription   string        `yaml:"longDescription,omitempty"`
 	Example           string        `yaml:"example,omitempty"`
@@ -158,4 +160,12 @@ func LoadConfig(file string) (*Config, error) {
 	log.Tracef("definition: %+v", config)
 
 	return &config, nil
+}
+
+func (c *PackageConfig) GetDestinationPath() string {
+	if c.Path != "" {
+		return c.Path
+	}
+
+	return "./"
 }
